requirements.txt
tldextract
requests
beautifulsoup4
lxml
matplotlib
baidu-aip